Removed policy sending from additional queryString parameters
The latest ADAL correctly sets the policy details of the query string
it generates in line with OAUTH standards. By passing through both the
policy parameter and the policy converted to a dictionary as “params”
the policy details were being added to the authorisation URL twice
causing a 404 in basic operation.

This also mean the +convertPolicyToDictionary function could be removed
as it is no longer required.